Added driver interface for rfm95/SX127x modules
The interface and code make use of the LoraLib library to interface with these devices. I also modified the library so that I could ensure OOK modulation after initializing the library with FSK. This interface should be enough for low level abstraction of data communication using these radio modules. The same interface shall be implemented for rfm69 and bluetooth in the near future.